+++ HISTORY.rst
@@ -11,12 +11,15 @@ End-User Summary
 
 - BAM statistics (including target coverage information) can now be imported and displayed.
 - Mitochondrial variants can now be properly displayed.
+- Added ``Delete Case`` button and functionality to case overview, only visible for superusers.
 
 Full Change List
 ================
 
 - Adding code for importing, storing, and displaying BAM quality control values.
 - Fixing ``urls`` configuration bug preventing chrMT matches.
+- Added ``Delete Case`` button and functionality to case overview, only visible for superusers.
+  Deletes record from ``Case`` and variants from ``SmallVariant``, ``StructuralVariant`` and ``StructuralVariantGeneAnnotation`` associated with this case.
 
 -------
 v0.17.4